Move images out of /static
