+++ src/images/faqHeaderBlob.png
