+++ src/images/voteBlob.png
