+++ src/images/votersEdge.png
