+++ src/pages/index.js
@@ -12,16 +12,16 @@ import SnapshotItem from "../components/snapshotItem"
 import CandidateItem from "../components/candidateItem"
 import BehindTheScenesItem from "../components/behindTheScenesItem"
 // Images
-import headerBlob from "./../../static/images/headerBlob.png"
-import tertiary from "./../../static/images/Tertiary.png"
-import BlankProfile from "./../../static/images/blankProfile.png"
-import blue from "./../../static/images/blue.png"
-import orange from "./../../static/images/orange.png"
-import green from "./../../static/images/green.png"
-import aboutBlob from "./../../static/images/aboutBlob.png"
-import learnMore from "./../../static/images/learnMore.png"
-import voteBlob from "./../../static/images/voteBlob.png"
-import registerToVote from "./../../static/images/registerToVote.png"
+import headerBlob from "../images/headerBlob.png"
+import tertiary from "../images/Tertiary.png"
+import BlankProfile from "../images/blankProfile.png"
+import blue from "../images/blue.png"
+import orange from "../images/orange.png"
+import green from "../images/green.png"
+import aboutBlob from "../images/aboutBlob.png"
+import learnMore from "../images/learnMore.png"
+import voteBlob from "../images/voteBlob.png"
+import registerToVote from "../images/registerToVote.png"
 import useWindowIsLarge from "../common/hooks/useWindowIsLarge"
 import { formatPercent } from '../common/util/formatters'
